requirements.txt
discord.py
aiofile
jishaku
